requirements.txt
PyQt6
pillow
cnocr
numpy
pygame
scikit-learn
requests
Cython
PyMuPDF
python-docx